+++ src/Process/ProcessInterface.php
@@ -64,7 +64,10 @@ interface ProcessInterface
 
     public function isSpreadErrorToApplicationExitCode(): bool;
 
-    /** @param array<mixed> $env */
+    /**
+     * @param (callable('err'|'out', string): void)|null $callback
+     * @param array<mixed> $env
+     */
     public function start(?callable $callback = null, array $env = []): void;
 
     public function stop(float $timeout = 10, ?int $signal = null): ?int;